SubprocessComponent --> PythonScriptComponent
