+++ python_modules/dagster/dagster/components/lib/executable_component/script_utils.py
@@ -1,22 +1,14 @@
 import os
 import shlex
-import shutil
 from collections.abc import Sequence
-from functools import cached_property
 from pathlib import Path
 from typing import Literal, Optional, Union
-
-from dagster_shared import check
 
 from dagster._core.execution.context.asset_check_execution_context import AssetCheckExecutionContext
 from dagster._core.execution.context.asset_execution_context import AssetExecutionContext
 from dagster._core.pipes.context import PipesExecutionResult
 from dagster._core.pipes.subprocess import PipesSubprocessClient
-from dagster.components.core.context import ComponentLoadContext
-from dagster.components.lib.executable_component.component import (
-    ExecutableComponent,
-    OpMetadataSpec,
-)
+from dagster.components.lib.executable_component.component import OpMetadataSpec
 
 
 class ScriptSpec(OpMetadataSpec):
@@ -57,30 +49,3 @@ def get_cmd(script_runner_exe: list[str], spec: ScriptSpec, path: str) -> list[s
         return [*script_runner_exe, abs_path, *spec.args]
     else:
         return [*script_runner_exe, abs_path]
-
-
-class SubprocessComponent(ExecutableComponent):
-    execution: ScriptSpec
-
-    @property
-    def op_metadata_spec(self) -> OpMetadataSpec:
-        return self._subprocess_spec
-
-    @cached_property
-    def _subprocess_spec(self) -> ScriptSpec:
-        return ScriptSpec.with_script_stem_as_default_name(self.execution)
-
-    def invoke_execute_fn(
-        self,
-        context: Union[AssetExecutionContext, AssetCheckExecutionContext],
-        component_load_context: ComponentLoadContext,
-    ) -> Sequence[PipesExecutionResult]:
-        assert not self.resource_keys, "Pipes subprocess scripts cannot have resources"
-        return invoke_runner(
-            context=context,
-            command=get_cmd(
-                script_runner_exe=[check.not_none(shutil.which("python"), "python not found")],
-                spec=self.execution,
-                path=str(component_load_context.path),
-            ),
-        )
