+++ python_modules/dagster/dagster_tests/components_tests/executable_component_tests/test_pipes_subprocess.py
@@ -6,9 +6,9 @@ from dagster._core.definitions.asset_key import AssetCheckKey, AssetKey
 from dagster._core.definitions.asset_selection import AssetSelection
 from dagster._core.definitions.materialize import materialize
 from dagster._core.definitions.metadata.metadata_value import TextMetadataValue
-from dagster.components.lib.executable_component.subprocess_component import (
+from dagster.components.lib.executable_component.python_script_component import (
+    PythonScriptComponent,
     ScriptSpec,
-    SubprocessComponent,
 )
 from dagster.components.testing import scaffold_defs_sandbox
 
@@ -21,13 +21,13 @@ def get_function_content(fn: Callable) -> str:
 
 
 def test_pipes_subprocess_script_hello_world() -> None:
-    with scaffold_defs_sandbox(component_cls=SubprocessComponent) as sandbox:
+    with scaffold_defs_sandbox(component_cls=PythonScriptComponent) as sandbox:
         execute_path = sandbox.defs_folder_path / "script.py"
         execute_path.write_text("print('hello world')")
 
         with sandbox.load(
             component_body={
-                "type": "dagster.components.lib.executable_component.subprocess_component.SubprocessComponent",
+                "type": "dagster.components.lib.executable_component.python_script_component.PythonScriptComponent",
                 "attributes": {
                     "execution": {
                         "name": "op_name",
@@ -41,7 +41,7 @@ def test_pipes_subprocess_script_hello_world() -> None:
                 },
             }
         ) as (component, defs):
-            assert isinstance(component, SubprocessComponent)
+            assert isinstance(component, PythonScriptComponent)
             assert isinstance(component.execution, ScriptSpec)
             assets_def = defs.get_assets_def("asset")
             result = materialize([assets_def])
@@ -58,7 +58,7 @@ def test_pipes_subprocess_script_with_custom_materialize_result() -> None:
             with open_dagster_pipes() as context:
                 context.report_asset_materialization(metadata={"foo": "bar"})
 
-    with scaffold_defs_sandbox(component_cls=SubprocessComponent) as sandbox:
+    with scaffold_defs_sandbox(component_cls=PythonScriptComponent) as sandbox:
         raw_source = inspect.getsource(code_to_copy)
         raw_source = "\n".join(raw_source.split("\n")[1:])
         raw_source = dedent(raw_source)
@@ -68,7 +68,7 @@ def test_pipes_subprocess_script_with_custom_materialize_result() -> None:
 
         with sandbox.load(
             component_body={
-                "type": "dagster.components.lib.executable_component.subprocess_component.SubprocessComponent",
+                "type": "dagster.components.lib.executable_component.python_script_component.PythonScriptComponent",
                 "attributes": {
                     "execution": {
                         "path": "op_name.py",
@@ -81,7 +81,7 @@ def test_pipes_subprocess_script_with_custom_materialize_result() -> None:
                 },
             }
         ) as (component, defs):
-            assert isinstance(component, SubprocessComponent)
+            assert isinstance(component, PythonScriptComponent)
             assert isinstance(component.execution, ScriptSpec)
             assets_def = defs.get_assets_def("asset")
             result = materialize([assets_def])
@@ -93,10 +93,10 @@ def test_pipes_subprocess_script_with_custom_materialize_result() -> None:
 
 
 def test_pipes_subprocess_script_with_name_override() -> None:
-    with scaffold_defs_sandbox(component_cls=SubprocessComponent) as sandbox:
+    with scaffold_defs_sandbox(component_cls=PythonScriptComponent) as sandbox:
         with sandbox.load(
             component_body={
-                "type": "dagster.components.lib.executable_component.subprocess_component.SubprocessComponent",
+                "type": "dagster.components.lib.executable_component.python_script_component.PythonScriptComponent",
                 "attributes": {
                     "execution": {
                         "name": "op_name_override",
@@ -125,13 +125,13 @@ def test_pipes_subprocess_script_with_checks_only() -> None:
                     passed=True,
                 )
 
-    with scaffold_defs_sandbox(component_cls=SubprocessComponent) as sandbox:
+    with scaffold_defs_sandbox(component_cls=PythonScriptComponent) as sandbox:
         execute_path = sandbox.defs_folder_path / "only_checks.py"
         execute_path.write_text(get_function_content(code_to_copy))
 
         with sandbox.load(
             component_body={
-                "type": "dagster.components.lib.executable_component.subprocess_component.SubprocessComponent",
+                "type": "dagster.components.lib.executable_component.python_script_component.PythonScriptComponent",
                 "attributes": {
                     "execution": {
                         "path": "only_checks.py",
@@ -145,7 +145,7 @@ def test_pipes_subprocess_script_with_checks_only() -> None:
                 },
             }
         ) as (component, defs):
-            assert isinstance(component, SubprocessComponent)
+            assert isinstance(component, PythonScriptComponent)
             assert isinstance(component.execution, ScriptSpec)
             asset_check_key = AssetCheckKey(AssetKey("asset"), "check_name")
             check_def = defs.get_asset_checks_def(asset_check_key)
@@ -177,14 +177,14 @@ def test_pipes_subprocess_with_args() -> None:
         def arg_list():
             return ["arg_value"]
 
-    with scaffold_defs_sandbox(component_cls=SubprocessComponent) as sandbox:
+    with scaffold_defs_sandbox(component_cls=PythonScriptComponent) as sandbox:
         execute_path = sandbox.defs_folder_path / "op_name.py"
         execute_path.write_text(get_function_content(op_name_contents))
         template_vars_path = sandbox.defs_folder_path / "template_vars.py"
         template_vars_path.write_text(get_function_content(template_vars_content))
         with sandbox.load(
             component_body={
-                "type": "dagster.components.lib.executable_component.subprocess_component.SubprocessComponent",
+                "type": "dagster.components.lib.executable_component.python_script_component.PythonScriptComponent",
                 "attributes": {
                     "execution": {
                         "path": "op_name.py",
@@ -199,7 +199,7 @@ def test_pipes_subprocess_with_args() -> None:
                 "template_vars_module": ".template_vars",
             }
         ) as (component, defs):
-            assert isinstance(component, SubprocessComponent)
+            assert isinstance(component, PythonScriptComponent)
             assert isinstance(component.execution, ScriptSpec)
             assert component.execution.args == ["arg_value"]
             assets_def = defs.get_assets_def("asset")
@@ -220,12 +220,12 @@ def test_pipes_subprocess_with_inline_str() -> None:
             with open_dagster_pipes() as context:
                 context.report_asset_materialization(metadata={"arg": sys.argv[1]})
 
-    with scaffold_defs_sandbox(component_cls=SubprocessComponent) as sandbox:
+    with scaffold_defs_sandbox(component_cls=PythonScriptComponent) as sandbox:
         execute_path = sandbox.defs_folder_path / "op_name.py"
         execute_path.write_text(get_function_content(op_name_contents))
         with sandbox.load(
             component_body={
-                "type": "dagster.components.lib.executable_component.subprocess_component.SubprocessComponent",
+                "type": "dagster.components.lib.executable_component.python_script_component.PythonScriptComponent",
                 "attributes": {
                     "execution": {
                         "path": "op_name.py",
@@ -239,7 +239,7 @@ def test_pipes_subprocess_with_inline_str() -> None:
                 },
             }
         ) as (component, defs):
-            assert isinstance(component, SubprocessComponent)
+            assert isinstance(component, PythonScriptComponent)
             assert isinstance(component.execution, ScriptSpec)
             assert component.execution.args == "arg_value"
             assets_def = defs.get_assets_def("asset")
